requirements.txt
appnope==0.1.2
asgiref==3.3.1
autopep8==1.5.4
backcall==0.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.4
django-allauth==0.44.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
idna==2.10
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
oauthlib==3.1.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.7.3
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
traitlets==4.3.3
urllib3==1.26.2
wcwidth==0.2.5
